rename uwb wahlhandlung view
