+++ wls-gui-wahllokalsystem/src/plugins/router.ts
@@ -19,8 +19,8 @@ import EreignisseView from "@/views/EreignisseView.vue";
 import ExampleError404View from "@/views/ExampleError404View.vue";
 import HomeView from "@/views/HomeView.vue";
 import BWBWahlbriefErfassungView from "@/views/wahlhandlung/BWBWahlbriefErfassungView.vue";
+import UWBStimmabgabeView from "@/views/wahlhandlung/UWBStimmabgabeView.vue";
 import UWBWaehlerverzeichnisView from "@/views/wahlhandlung/UWBWaehlerverzeichnisView.vue";
-import UWBWahlhandlungView from "@/views/wahlhandlung/UWBWahlhandlungView.vue";
 import WahlbriefZulassungView from "@/views/wahlhandlung/WahlbriefZulassungView.vue";
 import WahleroeffnungView from "@/views/wahlhandlung/WahleroeffnungView.vue";
 import WahlumgebungView from "@/views/wahlhandlung/WahlumgebungView.vue";
@@ -52,7 +52,7 @@ const routes = [
   {
     path: "/wahlschliessung",
     name: ROUTE_WAHLSCHLIESSUNG,
-    component: UWBWahlhandlungView,
+    component: UWBStimmabgabeView,
     beforeEnter: permitNavigationOnlyForWahlbezirksArtUwb,
   },
   {
