+++ wls-gui-wahllokalsystem/src/views/wahlhandlung/UWBStimmabgabeView.vue
